@@ -1,9 +1,15 @@
 package com.augustbonds.momentsofjoy
 
+import android.util.Log
+
 class Logger {
     companion object {
         fun info(message: String) {
-            println("AUGUST: " + message)
+            Log.i("MomentsOfJoy", message)
+        }
+
+        fun debug(message: String) {
+            Log.d("MomentsOfJoy" ,message)
         }
     }
 }
